Fixes #2
A somewhat ugly, but seems-to-be-working, way of not determining the device ID when using `device_t` for the current device - while not forming a separate class for the current device. An alternative to this would be to split device_t into two separate classes, with the current device's class not keeping a `device_id` at all and obtaining it as necessary (e.g. for methods like `properties()`). It would mean defining almost every method we now have twice... but, on the other hand, the code will be clearer and less burdened with templating voodoo and compile-to-nothing statements. It would also probably mean we would need to wrap and re-throw exceptions to include infor
mation about the device ID - or forego that information for many `device_t` methods, which would be too bad. Quite the conundrum.