Fixes Issue #13.
The simpleStreams example was relying on the assumption that streams created with no specific flags synchronize implicitly with the default stream; that's valid when using cudaStreamCreate(), but was _not_ valid with the wrappers. While I think no implicit synchronization is the better default, I think I'll revert to the CUDA default behavior to avoid surprising developers.

Additional changes:

* Have rewritten some device-flags-setting code, it looked buggy
* Fixed a method overloading glitch which caused one variant of `device_t<...>::create_stream()` to never get called
* Comment and printed message tweaks in `simpleStreams.cu`