Fixes #40.
* Fixed: The collection of kernel launch argument addresses into an array was using an instantiation-frugal "trick", expanding all parameters at once - bu
t this trick did not guarantee that the order of evaluation would be the desired one. Have switched this to a more conservative recursive-instantion-based address collection function, for which the evaluation order _is_ guaranteed. I did reduce the number of instantiations to about one per 4 arguments rather than one-per-one, just in case.